backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.28.0
sqlalchemy>=2.1.0
pymysql>=1.1.0
cryptography>=42.0.0
pydantic>=2.6.0
python-dotenv>=1.0.1
python-dateutil>=2.8.2
